+++ WalkfulWidget/WalkfulWidgetBundle.swift
@@ -1,0 +1,9 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct WalkfulWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        StepWidget()
+    }
+}